Re-work page content loading to use remote function
